app/requirements.txt
flask==2.3.3
google-generativeai==0.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
types-PyYAML==6.0.12.12  # For type checking